Move types mocks to mocks folder
