+++ src/test/__mocks__/applicationTypes.js
