+++ src/test/addSourceWizard/addSourceWizard/sourceAddModal.test.js
@@ -3,8 +3,8 @@ import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import AddSourceWizard from '../../../components/addSourceWizard/SourceAddModal';
-import sourceTypes from '../helpers/sourceTypes';
-import applicationTypes from '../helpers/applicationTypes';
+import sourceTypes from '../../__mocks__/sourceTypes';
+import applicationTypes from '../../__mocks__/applicationTypes';
 
 import * as dependency from '../../../api/wizardHelpers';
 import render from '../__mocks__/render';
